Write function for splitting result into array
The result is now split into an array of single digits. They are strings
but this doesn't affect the function of the calculator overall, as it
works as long as the string is a single digit.

Also calling add commas in the event functions for when the equalsBtn
is pressed so that the result is formatted properly with commas.